added users field to category.. added testcases for testing users field

+++ pomlogger/views.py
@@ -97,10 +97,12 @@ def delete_entry(request,id):
     return redirect('pomlog_entry_archive_index')
 
 
-def _get_categories(catnames):
-    categories=string.split(catnames,',')
+def get_categories(catnames):
+    print 'get_categories():catnames=%s=%s'%(catnames,type(catnames))
+    #categories=string.split(catnames,',')
     cats=[]
-    for name in categories:
+    #for name in categories:
+    for name in catnames:
         if name and  not name.isspace():
             cat,status=PomCategory.objects.get_or_create(name__iexact=name.strip(),defaults={'name':name,'description':name})# if no defaults ,IntegrityError
             cats.append(cat)
@@ -115,8 +117,19 @@ def add_new_entry(request,template_name,page_title):
     context={'page_title':page_title,'entryform':form,'categoryform':catnameform}
     if request.method=='POST' and form.is_valid() and catnameform.is_valid():
         newentry=form.save()         
-        catnames=catnameform.cleaned_data['categories']            
-        newentry.categories=_get_categories(catnames)
+        catnames=catnameform.cleaned_data['categories']
+        catnames=get_list_of_names(catnames)            
+        #newentry.categories=get_categories(catnames)
+        #sj 2 mar 2010
+        cats=get_categories(catnames)
+        print 'add_new_entry():req.usr=',request.user
+        for x in cats:
+            print 'x.userscount=',x.users.count()
+            if request.user not in x.users.all():
+                print 'user:',request.user,'not in users'
+                x.users.add(request.user)#need to append, not assign
+            print 'x.userscount=',x.users.count()
+        newentry.categories=cats
         newentry.save()
         return redirect('pomlog_entry_archive_index')
         
@@ -125,29 +138,96 @@ def add_new_entry(request,template_name,page_title):
 def get_category_names_as_one_string(categorynameslist):
     return ','.join(categorynameslist)
 
+def get_list_of_names(names):
+    print 'get_list_of_names():names=',names,type(names)
+    return [x.strip() for x in names.split(',') if len( x.strip())  !=0]
+
+def add_user_to_categories(categories,user):
+    if categories:
+        for acat in categories:
+            if user not in acat.users.all():
+                acat.users.add(user)
+
+def remove_user_from_categories(categories,user):
+    if categories:
+        for acat in categories:
+            if user in acat.users.all():
+                acat.users.remove(user)
+
 @login_required
 @transaction.commit_on_success
 def edit_entry(request,id,template_name,page_title):
     entry=get_object_or_404(PomEntry,id=id,author=request.user)
-    categorynames_as_separate=[x.name for x in entry.categories.all()]
-    categorynames_as_one_string=get_category_names_as_one_string(categorynames_as_separate)
+    old_categorynames_list=[x.name for x in entry.categories.all()]    
+    categorynames_as_one_string=get_category_names_as_one_string(old_categorynames_list)
+    print 'before editing entry:categorynames_as_one_string=',categorynames_as_one_string
     categorynamesdata={'categories':categorynames_as_one_string}
     form_data=get_form_data(request)
     form=PomEntryForm(form_data,instance=entry)
     catnameform=PomCategoryNameForm(form_data,initial=categorynamesdata)
     context={'entryform':form,'categoryform':catnameform,'page_title':page_title}
-    if request.method=='POST' and form.is_valid() and catnameform.is_valid():       
+    form_valid=form.is_valid()
+    catnameform_valid=catnameform.is_valid()
+    print 'b4 post'
+    print 'form_valid=',form_valid
+    print 'catnameform_valid=',catnameform_valid 
+    if request.method=='POST' and form_valid and catnameform_valid:
+        print 'in post'       
         edited_entry=form.save()
         catnames=catnameform.cleaned_data['categories']
-        edited_entry.categories=_get_categories(catnames)
+        #catnames=get_list_of_names(catnames)
+        print 'after editing entry:catnames=',catnames
+        new_categorynames_list=get_list_of_names(catnames)
+        newlyaddedcatnames=set(new_categorynames_list)-set(old_categorynames_list)
+        newlyaddedcatnames=list(newlyaddedcatnames)
+        removedcatnames=set(old_categorynames_list)-set(new_categorynames_list)
+        removedcatnames=list(removedcatnames)
+        print 'edit_entry():newlyaddedcatnames=',newlyaddedcatnames
+        print 'edit_entry():removedcatnames=',removedcatnames
+        print 'edit_entry():catnames=',catnames
+        #for each name in newlyaddedcatnames,get or create category object, add this user in its users field
+        if newlyaddedcatnames:
+            newlyaddedcats=get_categories(newlyaddedcatnames)
+            add_user_to_categories(newlyaddedcats,request.user)
+        #for each name in removedcatnames ,get category object ,remove this user from its users field
+        if removedcatnames:
+            removedcats=get_categories(removedcatnames)
+            remove_user_from_categories(removedcats,request.user)
+        #sj 2 mar
+        cats=get_categories(get_list_of_names(catnames))
+        print 'edit_entry():got cats'
+        edited_entry.categories=cats
         edited_entry.save()           
         return redirect('pomlog_entry_archive_index')
     return custom_render(request,context,template_name)
 
+def update_cats_with_editable_status(user,categories):
+    cats={}
+    print '****************************'
+    for cat in categories:
+        usrcnt=cat.users.count()
+        allcatusrs=cat.users.all()
+        user_in_users=user in allcatusrs
+        print 'user',user,'in',allcatusrs,'=',user_in_users
+        if usrcnt==1 and user in allcatusrs:
+            print 'user :',user,' can edit :',cat 
+            cats[cat]=True
+        else:
+            print 'user :',user,' cannot edit :',cat 
+            cats[cat]=False
+    return cats
+            
+
 @login_required
 def category_list(request,template_name,page_title):
     categories=PomCategory.objects.all()
-    category_list_dict={'object_list':categories,'page_title':page_title }
+    cats_of_user=PomCategory.objects.filter(users=request.user)
+    print 'category_list():cats of :',request.user,' are=',cats_of_user
+    #category_list_dict={'object_list':categories,'page_title':page_title }
+    #create a dict with catobject as key and editablestatus as value
+    cats=update_cats_with_editable_status(request.user,cats_of_user)
+    print 'user:',request.user,cats
+    category_list_dict={'object_list':cats_of_user,'page_title':page_title ,'cats_status':cats}# user can only see his categories
     return custom_render(request,category_list_dict,template_name)
 
 @login_required
@@ -155,6 +235,7 @@ def category_detail(request,slug,template_name,page_title):
     category=get_object_or_404(PomCategory,slug=slug)
     now=datetime.datetime.now()
     context={'object':category,'now':now, 'page_title': page_title}
+    print 'category_detail():cat users=',category.users.all()
     return custom_render(request,context,template_name)
 
 @login_required
@@ -162,8 +243,11 @@ def delete_category(request,slug):
     print 'slug=',slug
     cat=get_object_or_404(PomCategory,slug=slug)
     print 'delete cat()=',cat
-    cat.delete()
-    print 'cat deleted'
+    if cat.users.count()==1 and request.user in cat.users.all():
+        cat.delete()
+        print 'cat deleted'
+    else:
+        print 'cannot delete category'       
     return redirect('pomlog_category_list')
